api/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.2
python-pkcs11==0.7.0
django-filter==21.1
whitenoise==5.3.0
cryptography==37.0.2
django-rest-knox==4.2.0 